requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Supabase
supabase==2.23.2
psycopg2-binary==2.9.11

# Document processing
pypdf==5.1.0
python-docx==1.1.2
beautifulsoup4==4.12.3
markdown==3.7

# LLM & Embeddings
anthropic==0.42.0
openai==1.59.8

# Utils
python-dotenv==1.0.1
